Separate SearchFragment and SearchProvider
