+++ app/src/main/java/org/jellyfin/androidtv/search/SearchFragment.java
@@ -1,7 +1,6 @@
 package org.jellyfin.androidtv.search;
 
 import android.os.Bundle;
-import android.os.Handler;
 import android.speech.SpeechRecognizer;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -11,25 +10,13 @@ import org.jellyfin.androidtv.R;
 import org.jellyfin.androidtv.itemhandling.BaseRowItem;
 import org.jellyfin.androidtv.itemhandling.ItemLauncher;
 import org.jellyfin.androidtv.itemhandling.ItemRowAdapter;
-import org.jellyfin.androidtv.util.Utils;
 
 import androidx.leanback.app.SearchSupportFragment;
-import androidx.leanback.widget.ArrayObjectAdapter;
 import androidx.leanback.widget.ListRow;
-import androidx.leanback.widget.ListRowPresenter;
-import androidx.leanback.widget.ObjectAdapter;
 import androidx.leanback.widget.SpeechOrbView;
 
-/**
- * Created by Eric on 1/26/2015.
- */
-public class SearchFragment extends SearchSupportFragment
-        implements SearchSupportFragment.SearchResultProvider {
+public class SearchFragment extends SearchSupportFragment {
 
-    private static final int SEARCH_DELAY_MS = 1500;
-    private final Handler mHandler = new Handler();
-    private ArrayObjectAdapter mRowsAdapter;
-    private SearchRunnable mDelayedLoad;
     private boolean isSpeechEnabled = false;
 
     @Override
@@ -37,14 +24,14 @@ public class SearchFragment extends SearchSupportFragment
         super.onCreate(savedInstanceState);
 
         isSpeechEnabled = SpeechRecognizer.isRecognitionAvailable(getContext());
-        mRowsAdapter = new ArrayObjectAdapter(new ListRowPresenter());
-        setSearchResultProvider(this);
+
+        SearchProvider searchProvider = new SearchProvider(getContext(), getActivity().getIntent().getBooleanExtra("MusicOnly", false));
+        setSearchResultProvider(searchProvider);
         setOnItemViewClickedListener((itemViewHolder, item, rowViewHolder, row) -> {
             if (!(item instanceof BaseRowItem)) return;
 
             ItemLauncher.launch((BaseRowItem) item, (ItemRowAdapter) ((ListRow) row).getAdapter(), ((BaseRowItem) item).getIndex(), getActivity());
         });
-        mDelayedLoad = new SearchRunnable(getActivity(), mRowsAdapter, getActivity().getIntent().getBooleanExtra("MusicOnly", false));
 
         // Disable speech functionality
         if (!isSpeechEnabled) {
@@ -68,49 +55,6 @@ public class SearchFragment extends SearchSupportFragment
         }
 
         return view;
-    }
-
-    @Override
-    public ObjectAdapter getResultsAdapter() {
-        return mRowsAdapter;
-    }
-
-    @Override
-    public boolean onQueryTextChange(String query) {
-        search(query, true);
-
-        return true;
-    }
-
-    @Override
-    public boolean onQueryTextSubmit(String query) {
-        search(query, false);
-
-        return true;
-    }
-
-    /**
-     * Update search results
-     *
-     * @param query   String to search for
-     * @param delayed When true the search is delayed by [SEARCH_DELAY_MS] milliseconds
-     */
-    private void search(String query, boolean delayed) {
-        // Clear results when query is empty
-        if (Utils.isEmpty(query)) {
-            mRowsAdapter.clear();
-            return;
-        }
-
-        // Remove current delayed search (if any)
-        mHandler.removeCallbacks(mDelayedLoad);
-
-        // Update search string
-        mDelayedLoad.setQueryString(query);
-
-        // Schedule search depending on [delayed]
-        if (delayed) mHandler.postDelayed(mDelayedLoad, SEARCH_DELAY_MS);
-        else mHandler.post(mDelayedLoad);
     }
 }
