+++ app/src/main/java/org/jellyfin/androidtv/search/SearchRunnable.java
@@ -1,22 +1,22 @@
 package org.jellyfin.androidtv.search;
 
-import android.app.Activity;
-import androidx.leanback.widget.ArrayObjectAdapter;
-import androidx.leanback.widget.HeaderItem;
-import androidx.leanback.widget.ListRow;
+import android.content.Context;
 
 import org.jellyfin.androidtv.R;
 import org.jellyfin.androidtv.itemhandling.ItemRowAdapter;
 import org.jellyfin.androidtv.presentation.CardPresenter;
-
 import org.jellyfin.apiclient.model.search.SearchQuery;
+
+import androidx.leanback.widget.ArrayObjectAdapter;
+import androidx.leanback.widget.HeaderItem;
+import androidx.leanback.widget.ListRow;
 
 /**
  * Created by Eric on 4/8/2015.
  */
 public class SearchRunnable implements Runnable {
     private String searchString;
-    private Activity activity;
+    private Context context;
     private ArrayObjectAdapter mRowsAdapter;
     private boolean musicOnly;
 
@@ -24,8 +24,8 @@ public class SearchRunnable implements Runnable {
         searchString = value;
     }
 
-    public SearchRunnable(Activity activity, ArrayObjectAdapter adapter, boolean musicOnly) {
-        this.activity = activity;
+    public SearchRunnable(Context context, ArrayObjectAdapter adapter, boolean musicOnly) {
+        this.context = context;
         this.mRowsAdapter = adapter;
         this.musicOnly = musicOnly;
     }
@@ -38,56 +38,56 @@ public class SearchRunnable implements Runnable {
             //Get search results by type
             SearchQuery people = getSearchQuery(new String[] {"Person","People"});
             ItemRowAdapter peopleAdapter = new ItemRowAdapter(people, new CardPresenter(), mRowsAdapter);
-            ListRow peopleRow = new ListRow(new HeaderItem(activity.getString(R.string.lbl_people)), peopleAdapter);
+            ListRow peopleRow = new ListRow(new HeaderItem(context.getString(R.string.lbl_people)), peopleAdapter);
             peopleAdapter.setRow(peopleRow);
             peopleAdapter.Retrieve();
 
             SearchQuery movies = getSearchQuery(new String[] {"Movie", "BoxSet"});
             ItemRowAdapter movieAdapter = new ItemRowAdapter(movies, new CardPresenter(), mRowsAdapter);
-            ListRow movieRow = new ListRow(new HeaderItem(activity.getString(R.string.lbl_movies)), movieAdapter);
+            ListRow movieRow = new ListRow(new HeaderItem(context.getString(R.string.lbl_movies)), movieAdapter);
             movieAdapter.setRow(movieRow);
             movieAdapter.Retrieve();
 
             SearchQuery tv = getSearchQuery(new String[] {"Series","Episode"});
             ItemRowAdapter tvAdapter = new ItemRowAdapter(tv, new CardPresenter(), mRowsAdapter);
-            ListRow tvRow = new ListRow(new HeaderItem(activity.getString(R.string.lbl_tv)), tvAdapter);
+            ListRow tvRow = new ListRow(new HeaderItem(context.getString(R.string.lbl_tv)), tvAdapter);
             tvAdapter.setRow(tvRow);
             tvAdapter.Retrieve();
 
             SearchQuery videos = getSearchQuery(new String[] {"Video"});
             ItemRowAdapter videoAdapter = new ItemRowAdapter(videos, new CardPresenter(), mRowsAdapter);
-            ListRow videoRow = new ListRow(new HeaderItem(activity.getString(R.string.lbl_videos)), videoAdapter);
+            ListRow videoRow = new ListRow(new HeaderItem(context.getString(R.string.lbl_videos)), videoAdapter);
             videoAdapter.setRow(videoRow);
             videoAdapter.Retrieve();
 
             SearchQuery recordings = getSearchQuery(new String[] {"Recording"});
             ItemRowAdapter recordingAdapter = new ItemRowAdapter(recordings, new CardPresenter(), mRowsAdapter);
-            ListRow recordingRow = new ListRow(new HeaderItem(activity.getString(R.string.lbl_recordings)), recordingAdapter);
+            ListRow recordingRow = new ListRow(new HeaderItem(context.getString(R.string.lbl_recordings)), recordingAdapter);
             recordingAdapter.setRow(recordingRow);
             recordingAdapter.Retrieve();
 
             SearchQuery programs = getSearchQuery(new String[] {"Program"});
             ItemRowAdapter programAdapter = new ItemRowAdapter(programs, new CardPresenter(), mRowsAdapter);
-            ListRow programRow = new ListRow(new HeaderItem(activity.getString(R.string.lbl_programs)), programAdapter);
+            ListRow programRow = new ListRow(new HeaderItem(context.getString(R.string.lbl_programs)), programAdapter);
             programAdapter.setRow(programRow);
             programAdapter.Retrieve();
         }
 
         SearchQuery artists = getSearchQuery(new String[] {"MusicArtist"});
         ItemRowAdapter artistAdapter = new ItemRowAdapter(artists, new CardPresenter(), mRowsAdapter);
-        ListRow artistRow = new ListRow(new HeaderItem(activity.getString(R.string.lbl_artists)), artistAdapter);
+        ListRow artistRow = new ListRow(new HeaderItem(context.getString(R.string.lbl_artists)), artistAdapter);
         artistAdapter.setRow(artistRow);
         artistAdapter.Retrieve();
 
         SearchQuery albums = getSearchQuery(new String[] {"MusicAlbum"});
         ItemRowAdapter albumAdapter = new ItemRowAdapter(albums, new CardPresenter(), mRowsAdapter);
-        ListRow albumRow = new ListRow(new HeaderItem(activity.getString(R.string.lbl_albums)), albumAdapter);
+        ListRow albumRow = new ListRow(new HeaderItem(context.getString(R.string.lbl_albums)), albumAdapter);
         albumAdapter.setRow(albumRow);
         albumAdapter.Retrieve();
 
         SearchQuery songs = getSearchQuery(new String[] {"Audio"});
         ItemRowAdapter songAdapter = new ItemRowAdapter(songs, new CardPresenter(), mRowsAdapter);
-        ListRow songRow = new ListRow(new HeaderItem(activity.getString(R.string.lbl_songs)), songAdapter);
+        ListRow songRow = new ListRow(new HeaderItem(context.getString(R.string.lbl_songs)), songAdapter);
         songAdapter.setRow(songRow);
         songAdapter.Retrieve();
